Include python in matrxi for cross version tests
