+++ mlflow/ml-package-versions.yml
@@ -9,12 +9,16 @@ sklearn:
     maximum: "1.4.1.post1"
     requirements:
       "< 1.0": ["scipy==1.7.3"]
+    python:
+      "== dev": "3.9"
     run: |
       pytest tests/sklearn/test_sklearn_model_export.py
 
   autologging:
     minimum: "0.24.1"
     maximum: "1.4.1.post1"
+    python:
+      "== dev": "3.9"
     requirements:
       ">= 0.0.0": ["matplotlib"]
       "< 1.0": ["scipy==1.7.3"]
@@ -83,6 +87,8 @@ keras:
   models:
     minimum: "3.0.2"
     maximum: "3.0.5"
+    python:
+      "== dev": "3.9"
     requirements:
       ">= 3.0.0": ["jax[cpu]>0.4"]
     run: |
@@ -92,6 +98,8 @@ keras:
   autologging:
     minimum: "3.0.2"
     maximum: "3.0.5"
+    python:
+      "== dev": "3.9"
     requirements:
       ">= 3.0.0": ["jax[cpu]>0.4"]
     run: |
@@ -107,6 +115,8 @@ tensorflow:
   models:
     minimum: "2.6.5"
     maximum: "2.15.0.post1"
+    python:
+      "== dev": "3.9"
     requirements:
       # Requirements to run tests for keras
       ">= 0.0.0": ["scikit-learn", "pyspark", "pyarrow", "transformers!=4.38.0,!=4.38.1"]
@@ -132,6 +142,8 @@ tensorflow:
   autologging:
     minimum: "2.6.5"
     maximum: "2.15.0.post1"
+    python:
+      "== dev": "3.9"
     requirements:
       "== dev": ["scikit-learn"]
       "< 2.7.0": ["pandas==1.3.5"]
@@ -321,6 +333,8 @@ statsmodels:
   models:
     minimum: "0.11.1"
     maximum: "0.14.1"
+    python:
+      "== dev": "3.9"
     requirements:
       "< 0.13.0": ["pandas==1.3.5", "scipy==1.7.3"]
     run: |
@@ -329,6 +343,8 @@ statsmodels:
   autologging:
     minimum: "0.11.1"
     maximum: "0.14.1"
+    python:
+      "== dev": "3.9"
     requirements:
       "< 0.13.0": ["pandas==1.3.5", "scipy==1.7.3"]
     run: |
